Disable the DownloadFromSRA test.
This workflow does not play well with the current testing framework.  At the moment, the bucket to which we write results deletes old data after three days.  This data can usually be restored by the nightly test via call-caching of results from the intermediate bucket.  Any workflow that uses the FinalizeTo* tasks does this.

DownloadFromSRA doesn't work this way.  It downloads files directly to a VM and immediately uploads them to the destination bucket, never storing data in an intermediate place.  Thus, when the destination output expires, there is no intermediate data storage from which call-caching can retreive the missing file.  This is why this test works for a few days after I rerun the test manually without call-caching, but then starts failing again a few days later.

The best solution for this problem is not yet clear, but this is a low-priority workflow.  We're just going to disable this test for the time being.