Basecalling improvements for Terra compatibility.
Because basecalling can be single-plex or multiplexed, its output can technically be 1-to-1 or 1-to-many.  That severely breaks compatibility with Terra's data model.  Thus, we now only return the directory to which the results are being written.  We will rely on scripts that scan input directories to populate the data model instead.